add aliases and detailed docs
