+++ docs/content/3.api/1.composables/prefetch-components.md
@@ -1,6 +1,11 @@
 # `prefetchComponents`
 
-Nuxt provides composables and utilities to give you fine-grained control over prefetching JavaScript in your client application.
+::StabilityEdge
+::
+
+Nuxt provides composables and utilities to give you fine-grained control over prefetching and preloading components.
+
+> Prefetching component downloads the code in the background, this is based on the assumption that the component will likely be used for rendering, enabling the component to load instantly if and when the user requests it. The component is downloaded and cached for anticipated future use without the user making an explicit request for it.
 
 You can use `prefetchComponents` to manually prefetch individual components that have been registered globally in your Nuxt app. (By default Nuxt registers these as async components.) You must use the Pascal-cased version of the component name.
 
@@ -11,5 +16,9 @@ await prefetchComponents(['MyGlobalComponent1', 'MyGlobalComponent2'])
 ```
 
 ::alert{icon=👉}
-On server, `prefetchComponents` will have no effect.
+Current implementation behaves exactly the same as [`preloadComponents`](/api/composables/preload-components) by preloading components instead of just prefetching we are working to improve this behavior.
+::
+
+::alert{icon=👉}
+Currently, on server, `prefetchComponents` will have no effect.
 ::
