+++ packages/nuxt/src/app/composables/index.ts
@@ -12,4 +12,4 @@ export type { CookieOptions, CookieRef } from './cookie'
 export { useRequestHeaders, useRequestEvent, setResponseStatus } from './ssr'
 export { abortNavigation, addRouteMiddleware, defineNuxtRouteMiddleware, navigateTo, useRoute, useActiveRoute, useRouter } from './router'
 export type { AddRouteMiddlewareOptions, RouteMiddleware } from './router'
-export { prefetchComponents } from './prefetch'
+export { preloadComponents, prefetchComponents } from './preload'
